requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.1.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
discord.py==1.4.1
gevent==20.6.2
greenlet==0.4.16
idna==2.10
markdownify==0.5.2
multidict==4.7.6
pycparser==2.20
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
websocket_client==0.57.0
websockets==8.1
yarl==1.5.1
zope.event==4.4
zope.interface==5.1.0
